python/requirements.txt
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
filelock==3.9.0
flake8==6.0.0
Flask==2.2.3
identify==2.5.18
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector-python==8.0.32
nodeenv==1.7.0
platformdirs==3.1.0
pre-commit==3.1.1
protobuf==3.20.3
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
PyYAML==6.0
virtualenv==20.20.0
Werkzeug==2.2.3
zipp==3.15.0
python-dotenv==1.0.0
